Don't inline newtype DFuns (#21229)
Our work on #21229 uncovered a way in which Specialise could produce unsound
specialisations, ultimately caused by reboxing of newtype dictionaries.

In the case of datatypes, we carefully avoid this through the mechanism in
`Note [ClassOp/DFun selection]` that refrains from inlining a DFun.
This patch uses the same mechanism for newtype dictionaries, thus upholding the
new `Note [Do not inline dictionary selectors]`.

Fixes #21229.