When compiling with -fprof-late make unfoldings stable after simplification.
We do this to prevent cost centres affecting cross-module inlining.

Consider code like this:

module A where
f = ...

module B where
g = ... f ...

If we insert a cost centre in f we still want g to see the code without
the cost centre in f because otherwise it might affect inlining decisions.