DmdAnal: Anticipate ANF in `dmdAnalStar` (#21257)
The observable change: Demand analysis now gives `x` in
``let x = ... in let y = x |> co in y `seq` fst y `seq` ()`` the same demand as
`y`, as criticised by #21257.
The details are in `Note [Anticipating ANF in demand analysis]`, which provides
justification for why `anticipateANF` needs to be so complex.

Fixes #21257.